:recycle: Refactor: Move formatters and pattern parsers to Foxtail::Intl namespace
This major refactoring reorganizes the CLDR formatting functionality to align with JavaScript Intl API conventions:

**Namespace Changes:**
- Foxtail::CLDR::Formatter::Number → Foxtail::Intl::NumberFormat
- Foxtail::CLDR::Formatter::DateTime → Foxtail::Intl::DateTimeFormat
- Foxtail::CLDR::Formatter::LocalTimezoneDetector → Foxtail::Intl::LocalTimezoneDetector
- Foxtail::CLDR::PatternParser::Number → Foxtail::Intl::PatternParser::Number
- Foxtail::CLDR::PatternParser::DateTime → Foxtail::Intl::PatternParser::DateTime

**File Structure:**
- Created new Foxtail::Intl module with proper Zeitwerk-compliant structure
- Moved all formatter and pattern parser implementations to lib/foxtail/intl/
- Updated corresponding test files to match new namespace
- Removed empty CLDR::Formatter and CLDR::PatternParser modules

**Implementation Updates:**
- Simplified namespace references within Intl module (e.g., PatternParser::Number vs Foxtail::Intl::PatternParser::Number)
- Updated all cross-references throughout the codebase
- Fixed require paths in compatibility testing tools
- Updated RuboCop configuration files to reflect new paths
- Updated documentation references to use new class names

**Benefits:**
- Better alignment with JavaScript Intl API naming conventions
- Cleaner namespace organization following modern Ruby patterns
- Improved developer experience with shorter relative references
- Zeitwerk-compliant file structure for autoloading

All tests pass and maintain full compatibility with existing functionality.

:robot: Generated with [Claude Code](https://claude.ai/code)
