+++ .rubocop_todo.yml
@@ -11,16 +11,16 @@
 # WaywardPredicates: nonzero?
 Naming/PredicateMethod:
   Exclude:
-    - 'lib/foxtail/cldr/formatter/date_time.rb'
+    - 'lib/foxtail/intl/date_time_format.rb'
 
 # Configuration parameters: EnforcedStyle, CheckMethodNames, CheckSymbols, AllowedIdentifiers, AllowedPatterns.
 # SupportedStyles: snake_case, normalcase, non_integer
 # AllowedIdentifiers: TLS1_1, TLS1_2, capture3, iso8601, rfc1123_date, rfc822, rfc2822, rfc3339, x86_64
 Naming/VariableNumber:
   Exclude:
-    - 'lib/foxtail/cldr/formatter/date_time.rb'
+    - 'lib/foxtail/intl/date_time_format.rb'
 
 # This cop supports unsafe autocorrection (--autocorrect-all).
 Style/CommentedKeyword:
   Exclude:
-    - 'lib/foxtail/cldr/formatter/date_time.rb'
+    - 'lib/foxtail/intl/date_time_format.rb'
