--- lib/foxtail/cldr/formatter.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module Foxtail
-  module CLDR
-    # CLDR-based formatters implementing Intl formatting functionality
-    module Formatter
-    end
-  end
-end
