+++ lib/foxtail/intl.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Foxtail
+  # CLDR-based Intl formatting functionality implementing JavaScript Intl API compatibility
+  module Intl
+  end
+end
