+++ spec/foxtail/intl/date_time_format_spec.rb
@@ -2,46 +2,46 @@
 
 require "time"
 
-RSpec.describe Foxtail::CLDR::Formatter::DateTime do
-  subject(:formatter) { Foxtail::CLDR::Formatter::DateTime.new(locale: locale("en")) }
+RSpec.describe Foxtail::Intl::DateTimeFormat do
+  subject(:formatter) { Foxtail::Intl::DateTimeFormat.new(locale: locale("en")) }
 
   let(:test_time) { Time.utc(2023, 6, 15, 14, 30, 45) }
 
   # Stub timezone detection to return JST for consistent test results
   before do
-    allow(Foxtail::CLDR::Formatter::LocalTimezoneDetector).to receive(:detect)
-      .and_return(instance_double(Foxtail::CLDR::Formatter::LocalTimezoneDetector::DetectedTimezone, id: "Asia/Tokyo"))
+    allow(Foxtail::Intl::LocalTimezoneDetector).to receive(:detect)
+      .and_return(instance_double(Foxtail::Intl::LocalTimezoneDetector::DetectedTimezone, id: "Asia/Tokyo"))
   end
 
   describe "#call" do
     context "with locale options" do
       it "formats with English locale" do
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: locale("en"), dateStyle: "full")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: locale("en"), dateStyle: "full")
         result = formatter.call(test_time)
         expect(result).to eq("Thursday, June 15, 2023")
       end
 
       it "formats with Japanese locale" do
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: locale("ja"), dateStyle: "full")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: locale("ja"), dateStyle: "full")
         result = formatter.call(test_time)
         expect(result).to eq("2023年6月15日木曜日")
       end
 
       it "formats month names in Japanese" do
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: locale("ja"), month: "long")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: locale("ja"), month: "long")
         result = formatter.call(test_time)
         expect(result).to eq("6月")
       end
 
       it "formats weekday names in Japanese" do
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: locale("ja"), weekday: "long")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: locale("ja"), weekday: "long")
         result = formatter.call(test_time)
         expect(result).to eq("木曜日")
       end
 
       it "raises CLDR::DataNotAvailable for unknown locales" do
         expect {
-          formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: locale("unknown"), month: "long")
+          formatter = Foxtail::Intl::DateTimeFormat.new(locale: locale("unknown"), month: "long")
           formatter.call(test_time)
         }.to raise_error(Foxtail::CLDR::Repository::DataNotAvailable)
       end
@@ -51,26 +51,26 @@ RSpec.describe Foxtail::CLDR::Formatter::DateTime do
       let(:en_locale) { locale("en") }
 
       it "formats with dateStyle medium" do
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, dateStyle: "medium")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, dateStyle: "medium")
         result = formatter.call(test_time)
         expect(result).to eq("Jun 15, 2023")
       end
 
       it "formats with dateStyle full" do
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, dateStyle: "full")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, dateStyle: "full")
         result = formatter.call(test_time)
         expect(result).to eq("Thursday, June 15, 2023")
       end
 
       it "formats with timeStyle short" do
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, timeStyle: "short")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, timeStyle: "short")
         result = formatter.call(test_time)
         # UTC 14:30 -> JST 23:30
         expect(result).to eq("11:30 PM")
       end
 
       it "combines dateStyle and timeStyle" do
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, dateStyle: "medium", timeStyle: "short")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, dateStyle: "medium", timeStyle: "short")
         result = formatter.call(test_time)
         # Updated to match current CLDR pattern format (UTC 14:30 -> JST 23:30)
         expect(result).to eq("Jun 15, 2023, 11:30 PM")
@@ -96,58 +96,58 @@ RSpec.describe Foxtail::CLDR::Formatter::DateTime do
       let(:ja_locale) { locale("ja") }
 
       it "formats with simple custom pattern" do
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, pattern: "yyyy-MM-dd")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, pattern: "yyyy-MM-dd")
         result = formatter.call(test_time)
         expect(result).to eq("2023-06-15")
       end
 
       it "formats with custom pattern including weekday and month names" do
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, pattern: "EEEE, MMMM d, yyyy")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, pattern: "EEEE, MMMM d, yyyy")
         result = formatter.call(test_time)
         expect(result).to eq("Thursday, June 15, 2023")
       end
 
       it "formats with abbreviated names in custom pattern" do
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, pattern: "EEE, MMM d")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, pattern: "EEE, MMM d")
         result = formatter.call(test_time)
         expect(result).to eq("Thu, Jun 15")
       end
 
       it "formats with time components in custom pattern" do
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, pattern: "HH:mm:ss")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, pattern: "HH:mm:ss")
         result = formatter.call(test_time)
         # UTC 14:30:45 -> JST 23:30:45
         expect(result).to eq("23:30:45")
       end
 
       it "formats with 12-hour time in custom pattern" do
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, pattern: "h:mm a")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, pattern: "h:mm a")
         result = formatter.call(test_time)
         # UTC 14:30 -> JST 23:30 (11:30 PM)
         expect(result).to eq("11:30 PM")
       end
 
       it "formats with complex custom pattern" do
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, pattern: "EEEE, d MMMM yyyy 'at' HH:mm")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, pattern: "EEEE, d MMMM yyyy 'at' HH:mm")
         result = formatter.call(test_time)
         # UTC 14:30 -> JST 23:30
         expect(result).to eq("Thursday, 15 June 2023 at 23:30")
       end
 
       it "uses locale-specific names in custom patterns" do
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: ja_locale, pattern: "yyyy年MMMMd日(EEEE)")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: ja_locale, pattern: "yyyy年MMMMd日(EEEE)")
         result = formatter.call(test_time)
         expect(result).to eq("2023年6月15日(木曜日)")
       end
 
       it "handles literal text in custom patterns" do
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, pattern: "'Today is' EEEE")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, pattern: "'Today is' EEEE")
         result = formatter.call(test_time)
         expect(result).to eq("Today is Thursday")
       end
 
       it "handles mixed tokens and literals" do
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, pattern: "'Date:' dd/MM/yyyy 'Time:' HH:mm")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, pattern: "'Date:' dd/MM/yyyy 'Time:' HH:mm")
         result = formatter.call(test_time)
         # UTC 14:30 -> JST 23:30
         expect(result).to eq("Date: 15/06/2023 Time: 23:30")
@@ -157,29 +157,29 @@ RSpec.describe Foxtail::CLDR::Formatter::DateTime do
         utc_time = Time.new(2023, 6, 15, 14, 30, 45, "+00:00")
 
         # VV: timezone ID
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, timeZone: "America/New_York", pattern: "VV")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, timeZone: "America/New_York", pattern: "VV")
         result = formatter.call(utc_time)
         expect(result).to eq("America/New_York")
 
         # VVV: exemplar city
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, timeZone: "America/New_York", pattern: "VVV")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, timeZone: "America/New_York", pattern: "VVV")
         result = formatter.call(utc_time)
         expect(result).to eq("New York") # Extracted from timezone ID
 
         # ZZZZZ: ISO offset format
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, timeZone: "America/New_York", pattern: "ZZZZZ")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, timeZone: "America/New_York", pattern: "ZZZZZ")
         result = formatter.call(utc_time)
         expect(result).to eq("-04:00") # EDT in June
 
         # Z: basic offset format
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, timeZone: "America/New_York", pattern: "Z")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, timeZone: "America/New_York", pattern: "Z")
         result = formatter.call(utc_time)
         expect(result).to eq("-0400") # EDT in June
       end
 
       it "formats complex pattern with timezone symbols" do
         utc_time = Time.new(2023, 6, 15, 14, 30, 45, "+00:00")
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, timeZone: "Asia/Tokyo", pattern: "yyyy-MM-dd HH:mm VV (ZZZZZ)")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, timeZone: "Asia/Tokyo", pattern: "yyyy-MM-dd HH:mm VV (ZZZZZ)")
         result = formatter.call(utc_time)
         expect(result).to eq("2023-06-15 23:30 Asia/Tokyo (+09:00)")
       end
@@ -187,7 +187,7 @@ RSpec.describe Foxtail::CLDR::Formatter::DateTime do
 
     context "with timeZone options" do
       it "formats with UTC timezone" do
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: locale("en"), timeZone: "UTC", pattern: "HH:mm")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: locale("en"), timeZone: "UTC", pattern: "HH:mm")
         result = formatter.call(test_time)
         # test_time is UTC 14:30, so with UTC timezone option it stays 14:30
         expect(result).to eq("14:30")
@@ -195,7 +195,7 @@ RSpec.describe Foxtail::CLDR::Formatter::DateTime do
 
       it "formats with offset timezone +09:00" do
         utc_time = Time.new(2023, 6, 15, 5, 30, 45, "+00:00")
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: locale("en"), timeZone: "+09:00", pattern: "HH:mm")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: locale("en"), timeZone: "+09:00", pattern: "HH:mm")
         result = formatter.call(utc_time)
         # 05:30 UTC + 9 hours = 14:30
         expect(result).to eq("14:30")
@@ -203,14 +203,14 @@ RSpec.describe Foxtail::CLDR::Formatter::DateTime do
 
       it "formats with offset timezone -05:00" do
         utc_time = Time.new(2023, 6, 15, 14, 30, 45, "+00:00")
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: locale("en"), timeZone: "-05:00", pattern: "HH:mm")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: locale("en"), timeZone: "-05:00", pattern: "HH:mm")
         result = formatter.call(utc_time)
         # 14:30 UTC - 5 hours = 09:30
         expect(result).to eq("09:30")
       end
 
       it "preserves original timezone when no timeZone option" do
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: locale("en"), pattern: "HH:mm")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: locale("en"), pattern: "HH:mm")
         result = formatter.call(test_time)
         # test_time is UTC 14:30, system timezone is Asia/Tokyo, so 14:30 UTC -> 23:30 JST
         expect(result).to eq("23:30")
@@ -219,7 +219,7 @@ RSpec.describe Foxtail::CLDR::Formatter::DateTime do
       it "formats with IANA timezone America/New_York" do
         # June is during DST, so EDT (UTC-4)
         utc_time = Time.new(2023, 6, 15, 14, 30, 45, "+00:00")
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: locale("en"), timeZone: "America/New_York", pattern: "HH:mm")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: locale("en"), timeZone: "America/New_York", pattern: "HH:mm")
         result = formatter.call(utc_time)
         # 14:30 UTC - 4 hours = 10:30 EDT
         expect(result).to eq("10:30")
@@ -227,7 +227,7 @@ RSpec.describe Foxtail::CLDR::Formatter::DateTime do
 
       it "formats with IANA timezone Asia/Tokyo" do
         utc_time = Time.new(2023, 6, 15, 5, 30, 45, "+00:00")
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: locale("en"), timeZone: "Asia/Tokyo", pattern: "HH:mm")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: locale("en"), timeZone: "Asia/Tokyo", pattern: "HH:mm")
         result = formatter.call(utc_time)
         # 05:30 UTC + 9 hours = 14:30 JST
         expect(result).to eq("14:30")
@@ -236,7 +236,7 @@ RSpec.describe Foxtail::CLDR::Formatter::DateTime do
       it "formats with IANA timezone Europe/London" do
         # June is during BST (British Summer Time, UTC+1)
         utc_time = Time.new(2023, 6, 15, 14, 30, 45, "+00:00")
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: locale("en"), timeZone: "Europe/London", pattern: "HH:mm")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: locale("en"), timeZone: "Europe/London", pattern: "HH:mm")
         result = formatter.call(utc_time)
         # 14:30 UTC + 1 hour = 15:30 BST
         expect(result).to eq("15:30")
@@ -245,7 +245,7 @@ RSpec.describe Foxtail::CLDR::Formatter::DateTime do
       it "raises error for invalid timezone" do
         utc_time = Time.new(2023, 6, 15, 14, 30, 45, "+00:00")
         expect {
-          formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: locale("en"), timeZone: "Invalid/Timezone", pattern: "HH:mm")
+          formatter = Foxtail::Intl::DateTimeFormat.new(locale: locale("en"), timeZone: "Invalid/Timezone", pattern: "HH:mm")
           formatter.call(utc_time)
         }.to raise_error(TZInfo::InvalidTimezoneIdentifier)
       end
@@ -254,19 +254,19 @@ RSpec.describe Foxtail::CLDR::Formatter::DateTime do
         utc_time = Time.new(2023, 6, 15, 14, 30, 45, "+00:00")
 
         # Test Etc/UTC timezone handling
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: locale("en"), timeZone: "Etc/UTC", pattern: "z")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: locale("en"), timeZone: "Etc/UTC", pattern: "z")
         result = formatter.call(utc_time)
         expect(result).to be_a(String)
         expect(result).not_to be_empty
 
         # Test offset format timezone
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: locale("en"), timeZone: "+09:00", pattern: "z")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: locale("en"), timeZone: "+09:00", pattern: "z")
         result = formatter.call(utc_time)
         expect(result).to be_a(String)
         expect(result).not_to be_empty
 
         # Test GMT metazone with different locales
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: locale("fr"), timeZone: "UTC", pattern: "z")
+        formatter = Foxtail::Intl::DateTimeFormat.new(locale: locale("fr"), timeZone: "UTC", pattern: "z")
         result = formatter.call(utc_time)
         expect(result).to be_a(String)
         expect(result).not_to be_empty
@@ -283,14 +283,14 @@ RSpec.describe Foxtail::CLDR::Formatter::DateTime do
       # Test hour field adaptation with different options
       describe "hour field formatting" do
         it "formats hour as numeric (1-digit for non-zero)" do
-          formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, hour: "numeric", minute: "2-digit", second: "2-digit", hour12: false, timeZone: "UTC")
+          formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, hour: "numeric", minute: "2-digit", second: "2-digit", hour12: false, timeZone: "UTC")
           result = formatter.call(midnight_utc)
           # Should follow CLDR pattern for English (HH -> keep as 2-digit for en-US)
           expect(result).to match(/\d{1,2}:00:00/)
         end
 
         it "formats hour as 2-digit (always padded)" do
-          formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, hour: "2-digit", minute: "2-digit", second: "2-digit", hour12: false, timeZone: "UTC")
+          formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, hour: "2-digit", minute: "2-digit", second: "2-digit", hour12: false, timeZone: "UTC")
           result = formatter.call(midnight_utc)
           expect(result).to eq("00:00:00")
         end
@@ -298,21 +298,21 @@ RSpec.describe Foxtail::CLDR::Formatter::DateTime do
         it "handles different locales with hour numeric" do
           # Test multiple locales to ensure pattern adaptation works
           [en_locale, de_locale, fr_locale, ja_locale].each do |test_locale|
-            formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: test_locale, hour: "numeric", minute: "2-digit", hour12: false, timeZone: "UTC")
+            formatter = Foxtail::Intl::DateTimeFormat.new(locale: test_locale, hour: "numeric", minute: "2-digit", hour12: false, timeZone: "UTC")
             result = formatter.call(midnight_utc)
             expect(result).to match(/\d{1,2}:00/) # Should have proper hour formatting
           end
         end
 
         it "handles 12-hour format with hour options" do
-          formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, hour: "numeric", minute: "2-digit", hour12: true, timeZone: "UTC")
+          formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, hour: "numeric", minute: "2-digit", hour12: true, timeZone: "UTC")
           result = formatter.call(midnight_utc)
           expect(result).to match(/\d{1,2}:00 (AM|PM)/)
         end
 
         it "handles non-zero hours with numeric formatting" do
           one_pm_utc = Time.new(2023, 12, 25, 13, 0, 0, "+00:00")
-          formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, hour: "numeric", minute: "2-digit", hour12: false, timeZone: "UTC")
+          formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, hour: "numeric", minute: "2-digit", hour12: false, timeZone: "UTC")
           result = formatter.call(one_pm_utc)
           expect(result).to match(/1?3:00/)
         end
@@ -322,14 +322,14 @@ RSpec.describe Foxtail::CLDR::Formatter::DateTime do
       describe "minute and second field formatting" do
         it "formats minute as 2-digit" do
           five_minutes_utc = Time.new(2023, 12, 25, 0, 5, 0, "+00:00")
-          formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, hour: "numeric", minute: "2-digit", hour12: false, timeZone: "UTC")
+          formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, hour: "numeric", minute: "2-digit", hour12: false, timeZone: "UTC")
           result = formatter.call(five_minutes_utc)
           expect(result).to match(/\d{1,2}:05/)
         end
 
         it "formats second as 2-digit" do
           five_seconds_utc = Time.new(2023, 12, 25, 0, 0, 5, "+00:00")
-          formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, hour: "numeric", minute: "2-digit", second: "2-digit", hour12: false, timeZone: "UTC")
+          formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, hour: "numeric", minute: "2-digit", second: "2-digit", hour12: false, timeZone: "UTC")
           result = formatter.call(five_seconds_utc)
           expect(result).to match(/\d{1,2}:00:05/)
         end
@@ -338,31 +338,31 @@ RSpec.describe Foxtail::CLDR::Formatter::DateTime do
       # Test single field formatting (triggers format_fields_individually)
       describe "single field formatting" do
         it "formats only weekday" do
-          formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, weekday: "long")
+          formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, weekday: "long")
           result = formatter.call(test_time)
           expect(result).to eq("Thursday")
         end
 
         it "formats only month" do
-          formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, month: "long")
+          formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, month: "long")
           result = formatter.call(test_time)
           expect(result).to eq("June")
         end
 
         it "formats only year" do
-          formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, year: "numeric")
+          formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, year: "numeric")
           result = formatter.call(test_time)
           expect(result).to eq("2023")
         end
 
         it "formats only day" do
-          formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, day: "numeric")
+          formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, day: "numeric")
           result = formatter.call(test_time)
           expect(result).to eq("15")
         end
 
         it "formats only hour" do
-          formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, hour: "numeric", hour12: false, timeZone: "UTC")
+          formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, hour: "numeric", hour12: false, timeZone: "UTC")
           result = formatter.call(test_time)
           expect(result).to eq("14")
         end
@@ -371,25 +371,25 @@ RSpec.describe Foxtail::CLDR::Formatter::DateTime do
       # Test complex field combinations that might trigger fallback patterns
       describe "complex field combinations" do
         it "handles hour, minute, second combination" do
-          formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, hour: "2-digit", minute: "2-digit", second: "2-digit", hour12: false, timeZone: "UTC")
+          formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, hour: "2-digit", minute: "2-digit", second: "2-digit", hour12: false, timeZone: "UTC")
           result = formatter.call(test_time)
           expect(result).to eq("14:30:45")
         end
 
         it "handles year, month, day combination" do
-          formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, year: "numeric", month: "long", day: "numeric")
+          formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, year: "numeric", month: "long", day: "numeric")
           result = formatter.call(test_time)
           expect(result).to eq("June 15, 2023")
         end
 
         it "handles weekday, year, month, day combination" do
-          formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, weekday: "long", year: "numeric", month: "short", day: "numeric")
+          formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, weekday: "long", year: "numeric", month: "short", day: "numeric")
           result = formatter.call(test_time)
           expect(result).to eq("Thursday, Jun 15, 2023")
         end
 
         it "handles mixed date and time fields" do
-          formatter = Foxtail::CLDR::Formatter::DateTime.new(
+          formatter = Foxtail::Intl::DateTimeFormat.new(
             locale: en_locale,
             year: "numeric",
             month: "short",
@@ -417,7 +417,7 @@ RSpec.describe Foxtail::CLDR::Formatter::DateTime do
           ]
 
           test_cases.each do |test_case|
-            formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, month: test_case[:width])
+            formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, month: test_case[:width])
             result = formatter.call(test_time)
             expect(result).to eq(test_case[:expected])
           end
@@ -430,7 +430,7 @@ RSpec.describe Foxtail::CLDR::Formatter::DateTime do
           ]
 
           test_cases.each do |test_case|
-            formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: en_locale, weekday: test_case[:width])
+            formatter = Foxtail::Intl::DateTimeFormat.new(locale: en_locale, weekday: test_case[:width])
             result = formatter.call(test_time)
             expect(result).to eq(test_case[:expected])
           end
@@ -451,7 +451,7 @@ RSpec.describe Foxtail::CLDR::Formatter::DateTime do
       # Currently Foxtail produces: 10:30:00 午前
 
       it "positions AM marker before time (morning)" do
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(
+        formatter = Foxtail::Intl::DateTimeFormat.new(
           locale: ja_locale,
           hour: "numeric",
           minute: "2-digit",
@@ -464,7 +464,7 @@ RSpec.describe Foxtail::CLDR::Formatter::DateTime do
       end
 
       it "positions AM marker before time (midnight)" do
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(
+        formatter = Foxtail::Intl::DateTimeFormat.new(
           locale: ja_locale,
           hour: "numeric",
           minute: "2-digit",
@@ -478,7 +478,7 @@ RSpec.describe Foxtail::CLDR::Formatter::DateTime do
       end
 
       it "positions PM marker before time (noon)" do
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(
+        formatter = Foxtail::Intl::DateTimeFormat.new(
           locale: ja_locale,
           hour: "numeric",
           minute: "2-digit",
@@ -492,7 +492,7 @@ RSpec.describe Foxtail::CLDR::Formatter::DateTime do
       end
 
       it "positions PM marker before time (evening)" do
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(
+        formatter = Foxtail::Intl::DateTimeFormat.new(
           locale: ja_locale,
           hour: "numeric",
           minute: "2-digit",
@@ -505,7 +505,7 @@ RSpec.describe Foxtail::CLDR::Formatter::DateTime do
       end
 
       it "positions PM marker before time (late night)" do
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(
+        formatter = Foxtail::Intl::DateTimeFormat.new(
           locale: ja_locale,
           hour: "numeric",
           minute: "2-digit",
@@ -521,42 +521,42 @@ RSpec.describe Foxtail::CLDR::Formatter::DateTime do
     context "with special values (Infinity, NaN)" do
       it "raises an error for positive infinity" do
         expect {
-          formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: locale("en"), dateStyle: "medium")
+          formatter = Foxtail::Intl::DateTimeFormat.new(locale: locale("en"), dateStyle: "medium")
           formatter.call(Float::INFINITY)
         }.to raise_error(ArgumentError, /special value|invalid.*time/i)
       end
 
       it "raises an error for negative infinity" do
         expect {
-          formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: locale("en"), dateStyle: "medium")
+          formatter = Foxtail::Intl::DateTimeFormat.new(locale: locale("en"), dateStyle: "medium")
           formatter.call(-Float::INFINITY)
         }.to raise_error(ArgumentError, /special value|invalid.*time/i)
       end
 
       it "raises an error for NaN" do
         expect {
-          formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: locale("en"), dateStyle: "medium")
+          formatter = Foxtail::Intl::DateTimeFormat.new(locale: locale("en"), dateStyle: "medium")
           formatter.call(Float::NAN)
         }.to raise_error(ArgumentError, /special value|invalid.*time/i)
       end
 
       it "raises an error for string representations of infinity" do
         expect {
-          formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: locale("en"), dateStyle: "medium")
+          formatter = Foxtail::Intl::DateTimeFormat.new(locale: locale("en"), dateStyle: "medium")
           formatter.call("Infinity")
         }.to raise_error(ArgumentError, /special value|invalid.*time/i)
       end
 
       it "raises an error for string representations of negative infinity" do
         expect {
-          formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: locale("en"), dateStyle: "medium")
+          formatter = Foxtail::Intl::DateTimeFormat.new(locale: locale("en"), dateStyle: "medium")
           formatter.call("-Infinity")
         }.to raise_error(ArgumentError, /special value|invalid.*time/i)
       end
 
       it "raises an error for string representations of NaN" do
         expect {
-          formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: locale("en"), dateStyle: "medium")
+          formatter = Foxtail::Intl::DateTimeFormat.new(locale: locale("en"), dateStyle: "medium")
           formatter.call("NaN")
         }.to raise_error(ArgumentError, /special value|invalid.*time/i)
       end
@@ -565,21 +565,21 @@ RSpec.describe Foxtail::CLDR::Formatter::DateTime do
         # String with 400 digits that will overflow to Infinity when converted to Float
         huge_number_string = "1#{"0" * 400}"
         expect {
-          formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: locale("en"), dateStyle: "medium")
+          formatter = Foxtail::Intl::DateTimeFormat.new(locale: locale("en"), dateStyle: "medium")
           formatter.call(huge_number_string)
         }.to raise_error(ArgumentError, /special value|overflow|invalid.*time/i)
       end
 
       it "raises an error for numeric strings with extremely large exponents" do
         expect {
-          formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: locale("en"), dateStyle: "medium")
+          formatter = Foxtail::Intl::DateTimeFormat.new(locale: locale("en"), dateStyle: "medium")
           formatter.call("1.0e309")
         }.to raise_error(ArgumentError, /special value|overflow|invalid.*time/i)
       end
 
       it "raises an error for numeric strings exceeding Float::MAX" do
         expect {
-          formatter = Foxtail::CLDR::Formatter::DateTime.new(locale: locale("en"), dateStyle: "medium")
+          formatter = Foxtail::Intl::DateTimeFormat.new(locale: locale("en"), dateStyle: "medium")
           formatter.call("1.8e308")
         }.to raise_error(ArgumentError, /special value|overflow|invalid.*time/i)
       end
@@ -592,7 +592,7 @@ RSpec.describe Foxtail::CLDR::Formatter::DateTime do
       let(:time) { Time.new(2023, 7, 4, 15, 45, 30, "+00:00") }
 
       it "uses Unicode minus sign U+2212 for negative offsets" do
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(
+        formatter = Foxtail::Intl::DateTimeFormat.new(
           locale: fr_locale,
           timeStyle: "long",
           timeZone: "America/New_York" # UTC-4 in summer
@@ -604,7 +604,7 @@ RSpec.describe Foxtail::CLDR::Formatter::DateTime do
       end
 
       it "uses ASCII plus sign for positive offsets" do
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(
+        formatter = Foxtail::Intl::DateTimeFormat.new(
           locale: fr_locale,
           timeStyle: "long",
           timeZone: "Asia/Tokyo" # UTC+9
@@ -619,7 +619,7 @@ RSpec.describe Foxtail::CLDR::Formatter::DateTime do
       let(:time) { Time.new(2023, 7, 4, 15, 45, 30, "+00:00") }
 
       it "uses ASCII hyphen for negative offsets" do
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(
+        formatter = Foxtail::Intl::DateTimeFormat.new(
           locale: en_locale,
           timeStyle: "long",
           timeZone: "+03:00" # Explicit offset timezone
@@ -631,7 +631,7 @@ RSpec.describe Foxtail::CLDR::Formatter::DateTime do
 
       it "displays correct GMT offset during daylight saving time" do
         summer_time = Time.new(2023, 7, 4, 15, 45, 30, "+00:00") # UTC
-        formatter = Foxtail::CLDR::Formatter::DateTime.new(
+        formatter = Foxtail::Intl::DateTimeFormat.new(
           locale: en_locale,
           timeStyle: "long",
           timeZone: "Europe/London"
