+++ spec/foxtail/intl/number_format_spec.rb
@@ -1,37 +1,37 @@
 # frozen_string_literal: true
 
-RSpec.describe Foxtail::CLDR::Formatter::Number do
-  subject(:formatter) { Foxtail::CLDR::Formatter::Number.new(locale: locale("en")) }
+RSpec.describe Foxtail::Intl::NumberFormat do
+  subject(:formatter) { Foxtail::Intl::NumberFormat.new(locale: locale("en")) }
 
   describe "#call" do
     context "with CLDR locale support" do
       it "formats with English locale" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("en"))
+        formatter = Foxtail::Intl::NumberFormat.new(locale: locale("en"))
         result = formatter.call(1234.5)
         expect(result).to eq("1,234.5")
       end
 
       it "formats with German locale using comma decimal separator" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("de"))
+        formatter = Foxtail::Intl::NumberFormat.new(locale: locale("de"))
         result = formatter.call(1234.5)
         expect(result).to eq("1.234,5")
       end
 
       it "formats with French locale" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("fr"))
+        formatter = Foxtail::Intl::NumberFormat.new(locale: locale("fr"))
         result = formatter.call(1234.5)
         expect(result).to eq("1\u202F234,5") # \u202F is narrow no-break space used by French CLDR
       end
 
       it "formats large numbers with grouping" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("en"))
+        formatter = Foxtail::Intl::NumberFormat.new(locale: locale("en"))
         result = formatter.call(1_234_567.89)
         expect(result).to eq("1,234,567.89")
       end
 
       it "raises CLDR::DataNotAvailable for unknown locales" do
         expect {
-          formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("unknown"))
+          formatter = Foxtail::Intl::NumberFormat.new(locale: locale("unknown"))
           formatter.call(1234.5)
         }.to raise_error(Foxtail::CLDR::Repository::DataNotAvailable)
       end
@@ -42,13 +42,13 @@ RSpec.describe Foxtail::CLDR::Formatter::Number do
       let(:en_locale) { locale("en") }
 
       it "formats as percentage with locale" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: de_locale, style: "percent")
+        formatter = Foxtail::Intl::NumberFormat.new(locale: de_locale, style: "percent")
         result = formatter.call(0.75)
         expect(result).to eq("75\u00A0%") # \u00A0 is non-breaking space
       end
 
       it "formats as currency" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: en_locale, style: "currency", currency: "€")
+        formatter = Foxtail::Intl::NumberFormat.new(locale: en_locale, style: "currency", currency: "€")
         result = formatter.call(1234.5)
         expect(result).to eq("€1,234.50")
       end
@@ -56,19 +56,19 @@ RSpec.describe Foxtail::CLDR::Formatter::Number do
 
     context "with combined CLDR and precision options" do
       it "formats with locale and minimum fraction digits" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("de"), minimumFractionDigits: 2)
+        formatter = Foxtail::Intl::NumberFormat.new(locale: locale("de"), minimumFractionDigits: 2)
         result = formatter.call(42)
         expect(result).to eq("42,00")
       end
 
       it "formats percentages with precision" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("en"), style: "percent", minimumFractionDigits: 1)
+        formatter = Foxtail::Intl::NumberFormat.new(locale: locale("en"), style: "percent", minimumFractionDigits: 1)
         result = formatter.call(0.125)
         expect(result).to eq("12.5%")
       end
 
       it "formats currency with precision" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(
+        formatter = Foxtail::Intl::NumberFormat.new(
           locale: locale("en"),
           style: "currency",
           currency: "$",
@@ -85,19 +85,19 @@ RSpec.describe Foxtail::CLDR::Formatter::Number do
 
       describe "USD formatting" do
         it "formats positive USD amounts with proper symbol and decimals" do
-          formatter = Foxtail::CLDR::Formatter::Number.new(locale: en_locale, style: "currency", currency: "USD")
+          formatter = Foxtail::Intl::NumberFormat.new(locale: en_locale, style: "currency", currency: "USD")
           result = formatter.call(1234.50)
           expect(result).to eq("$1,234.50")
         end
 
         it "formats negative USD amounts" do
-          formatter = Foxtail::CLDR::Formatter::Number.new(locale: en_locale, style: "currency", currency: "USD")
+          formatter = Foxtail::Intl::NumberFormat.new(locale: en_locale, style: "currency", currency: "USD")
           result = formatter.call(-1234.50)
           expect(result).to eq("-$1,234.50")
         end
 
         it "formats negative USD with accounting style" do
-          formatter = Foxtail::CLDR::Formatter::Number.new(
+          formatter = Foxtail::Intl::NumberFormat.new(
             locale: en_locale,
             style: "currency",
             currency: "USD",
@@ -108,13 +108,13 @@ RSpec.describe Foxtail::CLDR::Formatter::Number do
         end
 
         it "formats large amounts with proper grouping" do
-          formatter = Foxtail::CLDR::Formatter::Number.new(locale: en_locale, style: "currency", currency: "USD")
+          formatter = Foxtail::Intl::NumberFormat.new(locale: en_locale, style: "currency", currency: "USD")
           result = formatter.call(1_234_567.89)
           expect(result).to eq("$1,234,567.89")
         end
 
         it "formats small amounts correctly" do
-          formatter = Foxtail::CLDR::Formatter::Number.new(locale: en_locale, style: "currency", currency: "USD")
+          formatter = Foxtail::Intl::NumberFormat.new(locale: en_locale, style: "currency", currency: "USD")
           result = formatter.call(5.99)
           expect(result).to eq("$5.99")
         end
@@ -122,13 +122,13 @@ RSpec.describe Foxtail::CLDR::Formatter::Number do
 
       describe "JPY formatting" do
         it "formats JPY with no decimal places" do
-          formatter = Foxtail::CLDR::Formatter::Number.new(locale: en_locale, style: "currency", currency: "JPY")
+          formatter = Foxtail::Intl::NumberFormat.new(locale: en_locale, style: "currency", currency: "JPY")
           result = formatter.call(1234)
           expect(result).to eq("¥1,234")
         end
 
         it "formats fractional JPY by rounding to whole numbers" do
-          formatter = Foxtail::CLDR::Formatter::Number.new(locale: en_locale, style: "currency", currency: "JPY")
+          formatter = Foxtail::Intl::NumberFormat.new(locale: en_locale, style: "currency", currency: "JPY")
           result = formatter.call(1234.67)
           expect(result).to eq("¥1,235") # Should round to nearest whole number
         end
@@ -137,12 +137,12 @@ RSpec.describe Foxtail::CLDR::Formatter::Number do
       describe "Currency digits" do
         it "respects CLDR currency fraction digits for different currencies" do
           # USD should have 2 decimal places
-          formatter = Foxtail::CLDR::Formatter::Number.new(locale: en_locale, style: "currency", currency: "USD")
+          formatter = Foxtail::Intl::NumberFormat.new(locale: en_locale, style: "currency", currency: "USD")
           usd_result = formatter.call(100)
           expect(usd_result).to eq("$100.00")
 
           # JPY should have 0 decimal places
-          formatter = Foxtail::CLDR::Formatter::Number.new(locale: en_locale, style: "currency", currency: "JPY")
+          formatter = Foxtail::Intl::NumberFormat.new(locale: en_locale, style: "currency", currency: "JPY")
           jpy_result = formatter.call(100)
           expect(jpy_result).to eq("¥100")
         end
@@ -150,13 +150,13 @@ RSpec.describe Foxtail::CLDR::Formatter::Number do
 
       describe "Locale-specific formatting", :integration do
         it "uses Japanese yen symbol in Japanese locale" do
-          formatter = Foxtail::CLDR::Formatter::Number.new(locale: ja_locale, style: "currency", currency: "JPY")
+          formatter = Foxtail::Intl::NumberFormat.new(locale: ja_locale, style: "currency", currency: "JPY")
           result = formatter.call(1234)
           expect(result).to eq("￥1,234")
         end
 
         it "formats USD in Japanese locale" do
-          formatter = Foxtail::CLDR::Formatter::Number.new(locale: ja_locale, style: "currency", currency: "USD")
+          formatter = Foxtail::Intl::NumberFormat.new(locale: ja_locale, style: "currency", currency: "USD")
           result = formatter.call(1234.50)
           expect(result).to eq("$1,234.50")
         end
@@ -164,7 +164,7 @@ RSpec.describe Foxtail::CLDR::Formatter::Number do
 
       describe "Error handling" do
         it "falls back to currency code when symbol is not available" do
-          formatter = Foxtail::CLDR::Formatter::Number.new(locale: en_locale, style: "currency", currency: "XYZ")
+          formatter = Foxtail::Intl::NumberFormat.new(locale: en_locale, style: "currency", currency: "XYZ")
           result = formatter.call(100)
           expect(result).to eq("XYZ100.00")
         end
@@ -174,14 +174,14 @@ RSpec.describe Foxtail::CLDR::Formatter::Number do
     context "with invalid input" do
       it "raises ArgumentError for invalid numeric strings" do
         expect {
-          formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("en"))
+          formatter = Foxtail::Intl::NumberFormat.new(locale: locale("en"))
           formatter.call("not a number")
         }.to raise_error(ArgumentError)
       end
 
       it "raises ArgumentError for unparseable numeric strings" do
         expect {
-          formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("en"))
+          formatter = Foxtail::Intl::NumberFormat.new(locale: locale("en"))
           formatter.call("12.34.56")
         }.to raise_error(ArgumentError)
       end
@@ -189,55 +189,55 @@ RSpec.describe Foxtail::CLDR::Formatter::Number do
 
     context "with special values" do
       it "formats positive infinity" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("en"))
+        formatter = Foxtail::Intl::NumberFormat.new(locale: locale("en"))
         result = formatter.call(Float::INFINITY)
         expect(result).to eq("∞")
       end
 
       it "formats negative infinity" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("en"))
+        formatter = Foxtail::Intl::NumberFormat.new(locale: locale("en"))
         result = formatter.call(-Float::INFINITY)
         expect(result).to eq("-∞")
       end
 
       it "formats NaN" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("en"))
+        formatter = Foxtail::Intl::NumberFormat.new(locale: locale("en"))
         result = formatter.call(Float::NAN)
         expect(result).to eq("NaN")
       end
 
       it "formats infinity with percent style" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("en"), style: "percent")
+        formatter = Foxtail::Intl::NumberFormat.new(locale: locale("en"), style: "percent")
         result = formatter.call(Float::INFINITY)
         expect(result).to eq("∞%")
       end
 
       it "formats negative infinity with currency style" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("en"), style: "currency", currency: "USD")
+        formatter = Foxtail::Intl::NumberFormat.new(locale: locale("en"), style: "currency", currency: "USD")
         result = formatter.call(-Float::INFINITY)
         expect(result).to eq("-$∞")
       end
 
       it "formats NaN with unit style" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("en"), style: "unit", unit: "meter", unitDisplay: "short")
+        formatter = Foxtail::Intl::NumberFormat.new(locale: locale("en"), style: "unit", unit: "meter", unitDisplay: "short")
         result = formatter.call(Float::NAN)
         expect(result).to eq("NaN m")
       end
 
       it "formats infinity with scientific notation" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("en"), notation: "scientific")
+        formatter = Foxtail::Intl::NumberFormat.new(locale: locale("en"), notation: "scientific")
         result = formatter.call(Float::INFINITY)
         expect(result).to eq("∞")
       end
 
       it "formats negative infinity with engineering notation" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("en"), notation: "engineering")
+        formatter = Foxtail::Intl::NumberFormat.new(locale: locale("en"), notation: "engineering")
         result = formatter.call(-Float::INFINITY)
         expect(result).to eq("-∞")
       end
 
       it "formats NaN with compact notation" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("en"), notation: "compact")
+        formatter = Foxtail::Intl::NumberFormat.new(locale: locale("en"), notation: "compact")
         result = formatter.call(Float::NAN)
         expect(result).to eq("NaN")
       end
@@ -245,42 +245,42 @@ RSpec.describe Foxtail::CLDR::Formatter::Number do
 
     context "with scientific notation" do
       it "formats numbers in scientific notation with default precision" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("en"), notation: "scientific")
+        formatter = Foxtail::Intl::NumberFormat.new(locale: locale("en"), notation: "scientific")
         result = formatter.call(123.456)
         # Default maximumFractionDigits is 3 for scientific notation (Node.js Intl behavior)
         expect(result).to match(/^1\.235E\+?2$/)
       end
 
       it "formats large numbers in scientific notation with rounding" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("en"), notation: "scientific")
+        formatter = Foxtail::Intl::NumberFormat.new(locale: locale("en"), notation: "scientific")
         result = formatter.call(1_234_567.89)
         # Should round to 3 decimal places: 1.23456789 → 1.235
         expect(result).to match(/^1\.235E\+?6$/)
       end
 
       it "formats small numbers in scientific notation" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("en"), notation: "scientific")
+        formatter = Foxtail::Intl::NumberFormat.new(locale: locale("en"), notation: "scientific")
         result = formatter.call(0.000123)
         # Should produce format like "1.23E-4"
         expect(result).to match(/^1\.23E-4$/)
       end
 
       it "formats negative numbers in scientific notation with rounding" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("en"), notation: "scientific")
+        formatter = Foxtail::Intl::NumberFormat.new(locale: locale("en"), notation: "scientific")
         result = formatter.call(-987_654.321)
         # Should round to 3 decimal places: 9.87654321 → 9.877
         expect(result).to match(/^-9\.877E\+?5$/)
       end
 
       it "formats integer with minimum digits" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("en"), notation: "scientific")
+        formatter = Foxtail::Intl::NumberFormat.new(locale: locale("en"), notation: "scientific")
         result = formatter.call(1)
         # Should produce "1E0" for simple integer
         expect(result).to match(/^1E\+?0$/)
       end
 
       it "formats decimal with necessary precision" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("en"), notation: "scientific")
+        formatter = Foxtail::Intl::NumberFormat.new(locale: locale("en"), notation: "scientific")
         result = formatter.call(12.3)
         # Should produce "1.23E1" maintaining necessary precision
         expect(result).to match(/^1\.23E\+?1$/)
@@ -296,43 +296,43 @@ RSpec.describe Foxtail::CLDR::Formatter::Number do
 
     describe "custom pattern support" do
       it "formats with quirky decimal pattern" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("en"), pattern: "###,###,##0.000")
+        formatter = Foxtail::Intl::NumberFormat.new(locale: locale("en"), pattern: "###,###,##0.000")
         result = formatter.call(1234.567)
         expect(result).to eq("1,234.567")
       end
 
       it "formats with unusual currency placement" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("en"), pattern: "#,##0.00¤")
+        formatter = Foxtail::Intl::NumberFormat.new(locale: locale("en"), pattern: "#,##0.00¤")
         result = formatter.call(1234.56)
         expect(result).to eq("1,234.56$")
       end
 
       it "formats with multiple literal texts" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("en"), pattern: "'Score:' #0 'points'")
+        formatter = Foxtail::Intl::NumberFormat.new(locale: locale("en"), pattern: "'Score:' #0 'points'")
         result = formatter.call(42)
         expect(result).to eq("Score: 42 points")
       end
 
       it "formats with zero-padded scientific notation" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("en"), pattern: "000.000E000")
+        formatter = Foxtail::Intl::NumberFormat.new(locale: locale("en"), pattern: "000.000E000")
         result = formatter.call(1234)
         expect(result).to eq("1.234E003")
       end
 
       it "formats with permille and literal suffix" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("en"), pattern: "#0.00‰ 'rate'")
+        formatter = Foxtail::Intl::NumberFormat.new(locale: locale("en"), pattern: "#0.00‰ 'rate'")
         result = formatter.call(0.789)
         expect(result).to eq("789.00‰ rate")
       end
 
       it "formats with accounting-style negative pattern" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("en"), pattern: "#0.00;[#0.00]")
+        formatter = Foxtail::Intl::NumberFormat.new(locale: locale("en"), pattern: "#0.00;[#0.00]")
         negative_result = formatter.call(-456.78)
         expect(negative_result).to eq("[456.78]")
       end
 
       it "custom pattern takes precedence over style option" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("en"), pattern: "'Value:' #0.000", style: "percent")
+        formatter = Foxtail::Intl::NumberFormat.new(locale: locale("en"), pattern: "'Value:' #0.000", style: "percent")
         result = formatter.call(0.5)
         # Should format as decimal with literal, not as percent (50.000 indicates percent style was applied)
         expect(result).to eq("Value: 50.000")
@@ -340,7 +340,7 @@ RSpec.describe Foxtail::CLDR::Formatter::Number do
 
       it "raises error for patterns with conflicting percent and permille symbols" do
         expect {
-          formatter = Foxtail::CLDR::Formatter::Number.new(locale: locale("en"), pattern: "#0.0%‰")
+          formatter = Foxtail::Intl::NumberFormat.new(locale: locale("en"), pattern: "#0.0%‰")
           formatter.call(0.123)
         }.to raise_error(ArgumentError, /Pattern cannot contain both percent \(.*?\) and permille \(.*?\)/)
       end
@@ -351,65 +351,65 @@ RSpec.describe Foxtail::CLDR::Formatter::Number do
         let(:en_locale) { locale("en") }
 
         it "formats singular currency name for integer 1" do
-          formatter = Foxtail::CLDR::Formatter::Number.new(locale: en_locale, pattern: "¤¤¤ #,##0.00", currency: "USD")
+          formatter = Foxtail::Intl::NumberFormat.new(locale: en_locale, pattern: "¤¤¤ #,##0.00", currency: "USD")
           result = formatter.call(1)
           expect(result).to eq("US dollar 1.00")
         end
 
         it "formats plural currency name for 1.0 (CLDR-compliant)" do
-          formatter = Foxtail::CLDR::Formatter::Number.new(locale: en_locale, pattern: "¤¤¤ #,##0.00", currency: "USD")
+          formatter = Foxtail::Intl::NumberFormat.new(locale: en_locale, pattern: "¤¤¤ #,##0.00", currency: "USD")
           result = formatter.call(1.0)
           expect(result).to eq("US dollars 1.00")
         end
 
         it "formats plural currency name for 2" do
-          formatter = Foxtail::CLDR::Formatter::Number.new(locale: en_locale, pattern: "¤¤¤ #,##0.00", currency: "USD")
+          formatter = Foxtail::Intl::NumberFormat.new(locale: en_locale, pattern: "¤¤¤ #,##0.00", currency: "USD")
           result = formatter.call(2)
           expect(result).to eq("US dollars 2.00")
         end
 
         it "formats plural currency name for 0" do
-          formatter = Foxtail::CLDR::Formatter::Number.new(locale: en_locale, pattern: "¤¤¤ #,##0.00", currency: "USD")
+          formatter = Foxtail::Intl::NumberFormat.new(locale: en_locale, pattern: "¤¤¤ #,##0.00", currency: "USD")
           result = formatter.call(0)
           expect(result).to eq("US dollars 0.00")
         end
 
         it "formats plural currency name for decimal values" do
-          formatter = Foxtail::CLDR::Formatter::Number.new(locale: en_locale, pattern: "¤¤¤ #,##0.00", currency: "USD")
+          formatter = Foxtail::Intl::NumberFormat.new(locale: en_locale, pattern: "¤¤¤ #,##0.00", currency: "USD")
           result = formatter.call(1.5)
           expect(result).to eq("US dollars 1.50")
         end
 
         it "formats plural currency name for large numbers" do
-          formatter = Foxtail::CLDR::Formatter::Number.new(locale: en_locale, pattern: "¤¤¤ #,##0.00", currency: "USD")
+          formatter = Foxtail::Intl::NumberFormat.new(locale: en_locale, pattern: "¤¤¤ #,##0.00", currency: "USD")
           result = formatter.call(1000)
           expect(result).to eq("US dollars 1,000.00")
         end
 
         it "formats currency name with negative values" do
-          formatter = Foxtail::CLDR::Formatter::Number.new(locale: en_locale, pattern: "¤¤¤ #,##0.00", currency: "USD")
+          formatter = Foxtail::Intl::NumberFormat.new(locale: en_locale, pattern: "¤¤¤ #,##0.00", currency: "USD")
           result = formatter.call(-1)
           expect(result).to eq("US dollar -1.00")
         end
 
         it "handles unknown currency code XXX properly" do
-          formatter = Foxtail::CLDR::Formatter::Number.new(locale: en_locale, pattern: "¤¤¤ #,##0.00", currency: "XXX")
+          formatter = Foxtail::Intl::NumberFormat.new(locale: en_locale, pattern: "¤¤¤ #,##0.00", currency: "XXX")
           result = formatter.call(1)
           expect(result).to eq("(unknown unit of currency) 1.00")
         end
 
         it "falls back to currency code for truly unknown currencies" do
-          formatter = Foxtail::CLDR::Formatter::Number.new(locale: en_locale, pattern: "¤¤¤ #,##0.00", currency: "ZZZ")
+          formatter = Foxtail::Intl::NumberFormat.new(locale: en_locale, pattern: "¤¤¤ #,##0.00", currency: "ZZZ")
           result = formatter.call(1)
           expect(result).to eq("ZZZ 1.00")
         end
 
         it "formats different currency names properly" do
-          formatter = Foxtail::CLDR::Formatter::Number.new(locale: en_locale, pattern: "¤¤¤ #,##0.00", currency: "EUR")
+          formatter = Foxtail::Intl::NumberFormat.new(locale: en_locale, pattern: "¤¤¤ #,##0.00", currency: "EUR")
           result = formatter.call(1)
           expect(result).to eq("euro 1.00")
 
-          formatter = Foxtail::CLDR::Formatter::Number.new(locale: en_locale, pattern: "¤¤¤ #,##0.00", currency: "EUR")
+          formatter = Foxtail::Intl::NumberFormat.new(locale: en_locale, pattern: "¤¤¤ #,##0.00", currency: "EUR")
           result = formatter.call(2)
           expect(result).to eq("euros 2.00")
         end
@@ -419,13 +419,13 @@ RSpec.describe Foxtail::CLDR::Formatter::Number do
         let(:en_locale) { locale("en") }
 
         it "formats with currency name at the end" do
-          formatter = Foxtail::CLDR::Formatter::Number.new(locale: en_locale, pattern: "#,##0.00 ¤¤¤", currency: "USD")
+          formatter = Foxtail::Intl::NumberFormat.new(locale: en_locale, pattern: "#,##0.00 ¤¤¤", currency: "USD")
           result = formatter.call(1)
           expect(result).to eq("1.00 US dollar")
         end
 
         it "formats with currency name in parentheses" do
-          formatter = Foxtail::CLDR::Formatter::Number.new(locale: en_locale, pattern: "#,##0.00 '('¤¤¤')'", currency: "USD")
+          formatter = Foxtail::Intl::NumberFormat.new(locale: en_locale, pattern: "#,##0.00 '('¤¤¤')'", currency: "USD")
           result = formatter.call(2)
           expect(result).to eq("2.00 (US dollars)")
         end
@@ -439,7 +439,7 @@ RSpec.describe Foxtail::CLDR::Formatter::Number do
           # Test will depend on whether Polish CLDR data is available
           # and properly configured with plural currency names
           begin
-            formatter = Foxtail::CLDR::Formatter::Number.new(locale: pl_locale, pattern: "¤¤¤ #,##0.00", currency: "PLN")
+            formatter = Foxtail::Intl::NumberFormat.new(locale: pl_locale, pattern: "¤¤¤ #,##0.00", currency: "PLN")
             result = formatter.call(1)
             # Should use singular form
             expect(result).to match(/zł|PLN/)
@@ -454,37 +454,37 @@ RSpec.describe Foxtail::CLDR::Formatter::Number do
       let(:en_locale) { locale("en") }
 
       it "formats with unit style using default meter" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: en_locale, style: "unit")
+        formatter = Foxtail::Intl::NumberFormat.new(locale: en_locale, style: "unit")
         result = formatter.call(5)
         expect(result).to eq("5 m")
       end
 
       it "formats with specified unit" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: en_locale, style: "unit", unit: "kilometer")
+        formatter = Foxtail::Intl::NumberFormat.new(locale: en_locale, style: "unit", unit: "kilometer")
         result = formatter.call(100)
         expect(result).to eq("100 km")
       end
 
       it "formats with different unit display styles" do
         # Test short display (default)
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: en_locale, style: "unit", unit: "meter", unitDisplay: "short")
+        formatter = Foxtail::Intl::NumberFormat.new(locale: en_locale, style: "unit", unit: "meter", unitDisplay: "short")
         result_short = formatter.call(2)
         expect(result_short).to eq("2 m")
 
         # Test long display - should use plural form for 2
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: en_locale, style: "unit", unit: "meter", unitDisplay: "long")
+        formatter = Foxtail::Intl::NumberFormat.new(locale: en_locale, style: "unit", unit: "meter", unitDisplay: "long")
         result_long = formatter.call(2)
         expect(result_long).to eq("2 meters")
       end
 
       it "handles fractional numbers" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: en_locale, style: "unit", unit: "meter")
+        formatter = Foxtail::Intl::NumberFormat.new(locale: en_locale, style: "unit", unit: "meter")
         result = formatter.call(1.5)
         expect(result).to eq("1.5 m")
       end
 
       it "applies number formatting rules (grouping)" do
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: en_locale, style: "unit", unit: "meter")
+        formatter = Foxtail::Intl::NumberFormat.new(locale: en_locale, style: "unit", unit: "meter")
         result = formatter.call(1000)
         expect(result).to eq("1,000 m")
 
@@ -495,7 +495,7 @@ RSpec.describe Foxtail::CLDR::Formatter::Number do
       it "respects locale-specific number formatting" do
         # German: period for thousands, comma for decimal
         de_locale = locale("de")
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: de_locale, style: "unit", unit: "meter")
+        formatter = Foxtail::Intl::NumberFormat.new(locale: de_locale, style: "unit", unit: "meter")
         result = formatter.call(1000)
         expect(result).to eq("1.000 m")
 
@@ -504,7 +504,7 @@ RSpec.describe Foxtail::CLDR::Formatter::Number do
 
         # French: thin space for thousands, comma for decimal
         fr_locale = locale("fr")
-        formatter = Foxtail::CLDR::Formatter::Number.new(locale: fr_locale, style: "unit", unit: "meter")
+        formatter = Foxtail::Intl::NumberFormat.new(locale: fr_locale, style: "unit", unit: "meter")
         result = formatter.call(1000)
         expect(result).to eq("1\u{202F}000\u{202F}m")
 
@@ -512,8 +512,8 @@ RSpec.describe Foxtail::CLDR::Formatter::Number do
         expect(result).to eq("1\u{202F}234,5\u{202F}m")
 
         # Test short vs long unit display differences
-        short_formatter = Foxtail::CLDR::Formatter::Number.new(locale: fr_locale, style: "unit", unit: "meter", unitDisplay: "short")
-        long_formatter = Foxtail::CLDR::Formatter::Number.new(locale: fr_locale, style: "unit", unit: "meter", unitDisplay: "long")
+        short_formatter = Foxtail::Intl::NumberFormat.new(locale: fr_locale, style: "unit", unit: "meter", unitDisplay: "short")
+        long_formatter = Foxtail::Intl::NumberFormat.new(locale: fr_locale, style: "unit", unit: "meter", unitDisplay: "long")
 
         result_short = short_formatter.call(2)
         result_long = long_formatter.call(2)
