+++ spec/foxtail/intl/pattern_parser/date_time_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-RSpec.describe Foxtail::CLDR::PatternParser::DateTime do
-  subject(:parser) { Foxtail::CLDR::PatternParser::DateTime.new }
+RSpec.describe Foxtail::Intl::PatternParser::DateTime do
+  subject(:parser) { Foxtail::Intl::PatternParser::DateTime.new }
 
   describe "#parse" do
     context "with empty or nil patterns" do
@@ -18,7 +18,7 @@ RSpec.describe Foxtail::CLDR::PatternParser::DateTime do
       it "parses year patterns" do
         tokens = parser.parse("yyyy")
         expect(tokens).to have_attributes(size: 1)
-        expect(tokens.first).to be_a(Foxtail::CLDR::PatternParser::DateTime::FieldToken)
+        expect(tokens.first).to be_a(Foxtail::Intl::PatternParser::DateTime::FieldToken)
         expect(tokens.first.value).to eq("yyyy")
         expect(tokens.first.field_type).to eq(:year)
         expect(tokens.first.field_length).to eq(4)
@@ -27,7 +27,7 @@ RSpec.describe Foxtail::CLDR::PatternParser::DateTime do
       it "parses month patterns" do
         tokens = parser.parse("MMMM")
         expect(tokens).to have_attributes(size: 1)
-        expect(tokens.first).to be_a(Foxtail::CLDR::PatternParser::DateTime::FieldToken)
+        expect(tokens.first).to be_a(Foxtail::Intl::PatternParser::DateTime::FieldToken)
         expect(tokens.first.value).to eq("MMMM")
         expect(tokens.first.field_type).to eq(:month)
         expect(tokens.first.field_length).to eq(4)
@@ -36,7 +36,7 @@ RSpec.describe Foxtail::CLDR::PatternParser::DateTime do
       it "parses day patterns" do
         tokens = parser.parse("dd")
         expect(tokens).to have_attributes(size: 1)
-        expect(tokens.first).to be_a(Foxtail::CLDR::PatternParser::DateTime::FieldToken)
+        expect(tokens.first).to be_a(Foxtail::Intl::PatternParser::DateTime::FieldToken)
         expect(tokens.first.value).to eq("dd")
         expect(tokens.first.field_type).to eq(:day)
         expect(tokens.first.field_length).to eq(2)
@@ -45,7 +45,7 @@ RSpec.describe Foxtail::CLDR::PatternParser::DateTime do
       it "parses weekday patterns" do
         tokens = parser.parse("EEEE")
         expect(tokens).to have_attributes(size: 1)
-        expect(tokens.first).to be_a(Foxtail::CLDR::PatternParser::DateTime::FieldToken)
+        expect(tokens.first).to be_a(Foxtail::Intl::PatternParser::DateTime::FieldToken)
         expect(tokens.first.value).to eq("EEEE")
         expect(tokens.first.field_type).to eq(:weekday)
         expect(tokens.first.field_length).to eq(4)
@@ -54,7 +54,7 @@ RSpec.describe Foxtail::CLDR::PatternParser::DateTime do
       it "parses hour patterns" do
         tokens = parser.parse("HH")
         expect(tokens).to have_attributes(size: 1)
-        expect(tokens.first).to be_a(Foxtail::CLDR::PatternParser::DateTime::FieldToken)
+        expect(tokens.first).to be_a(Foxtail::Intl::PatternParser::DateTime::FieldToken)
         expect(tokens.first.value).to eq("HH")
         expect(tokens.first.field_type).to eq(:hour)
         expect(tokens.first.field_length).to eq(2)
@@ -63,7 +63,7 @@ RSpec.describe Foxtail::CLDR::PatternParser::DateTime do
       it "parses AM/PM patterns" do
         tokens = parser.parse("a")
         expect(tokens).to have_attributes(size: 1)
-        expect(tokens.first).to be_a(Foxtail::CLDR::PatternParser::DateTime::FieldToken)
+        expect(tokens.first).to be_a(Foxtail::Intl::PatternParser::DateTime::FieldToken)
         expect(tokens.first.value).to eq("a")
         expect(tokens.first.field_type).to eq(:am_pm)
         expect(tokens.first.field_length).to eq(1)
@@ -74,14 +74,14 @@ RSpec.describe Foxtail::CLDR::PatternParser::DateTime do
       it "parses simple literals" do
         tokens = parser.parse("Info:")
         expect(tokens).to have_attributes(size: 1)
-        expect(tokens.first).to be_a(Foxtail::CLDR::PatternParser::DateTime::LiteralToken)
+        expect(tokens.first).to be_a(Foxtail::Intl::PatternParser::DateTime::LiteralToken)
         expect(tokens.first.value).to eq("Info:")
       end
 
       it "combines consecutive literal characters" do
         tokens = parser.parse("Info: Output:")
         expect(tokens).to have_attributes(size: 1)
-        expect(tokens.first).to be_a(Foxtail::CLDR::PatternParser::DateTime::LiteralToken)
+        expect(tokens.first).to be_a(Foxtail::Intl::PatternParser::DateTime::LiteralToken)
         expect(tokens.first.value).to eq("Info: Output:")
       end
     end
@@ -90,7 +90,7 @@ RSpec.describe Foxtail::CLDR::PatternParser::DateTime do
       it "parses simple quoted text" do
         tokens = parser.parse("'Today is'")
         expect(tokens).to have_attributes(size: 1)
-        expect(tokens.first).to be_a(Foxtail::CLDR::PatternParser::DateTime::QuotedToken)
+        expect(tokens.first).to be_a(Foxtail::Intl::PatternParser::DateTime::QuotedToken)
         expect(tokens.first.value).to eq("'Today is'")
         expect(tokens.first.literal_text).to eq("Today is")
       end
@@ -98,14 +98,14 @@ RSpec.describe Foxtail::CLDR::PatternParser::DateTime do
       it "handles escaped quotes" do
         tokens = parser.parse("'Today''s date'")
         expect(tokens).to have_attributes(size: 1)
-        expect(tokens.first).to be_a(Foxtail::CLDR::PatternParser::DateTime::QuotedToken)
+        expect(tokens.first).to be_a(Foxtail::Intl::PatternParser::DateTime::QuotedToken)
         expect(tokens.first.literal_text).to eq("Today's date")
       end
 
       it "handles empty quoted strings" do
         tokens = parser.parse("''")
         expect(tokens).to have_attributes(size: 1)
-        expect(tokens.first).to be_a(Foxtail::CLDR::PatternParser::DateTime::QuotedToken)
+        expect(tokens.first).to be_a(Foxtail::Intl::PatternParser::DateTime::QuotedToken)
         expect(tokens.first.literal_text).to eq("")
       end
     end
@@ -115,25 +115,25 @@ RSpec.describe Foxtail::CLDR::PatternParser::DateTime do
         tokens = parser.parse("EEEE, MMMM d, yyyy")
         expect(tokens).to have_attributes(size: 7)
 
-        expect(tokens[0]).to be_a(Foxtail::CLDR::PatternParser::DateTime::FieldToken)
+        expect(tokens[0]).to be_a(Foxtail::Intl::PatternParser::DateTime::FieldToken)
         expect(tokens[0].value).to eq("EEEE")
 
-        expect(tokens[1]).to be_a(Foxtail::CLDR::PatternParser::DateTime::LiteralToken)
+        expect(tokens[1]).to be_a(Foxtail::Intl::PatternParser::DateTime::LiteralToken)
         expect(tokens[1].value).to eq(", ")
 
-        expect(tokens[2]).to be_a(Foxtail::CLDR::PatternParser::DateTime::FieldToken)
+        expect(tokens[2]).to be_a(Foxtail::Intl::PatternParser::DateTime::FieldToken)
         expect(tokens[2].value).to eq("MMMM")
 
-        expect(tokens[3]).to be_a(Foxtail::CLDR::PatternParser::DateTime::LiteralToken)
+        expect(tokens[3]).to be_a(Foxtail::Intl::PatternParser::DateTime::LiteralToken)
         expect(tokens[3].value).to eq(" ")
 
-        expect(tokens[4]).to be_a(Foxtail::CLDR::PatternParser::DateTime::FieldToken)
+        expect(tokens[4]).to be_a(Foxtail::Intl::PatternParser::DateTime::FieldToken)
         expect(tokens[4].value).to eq("d")
 
-        expect(tokens[5]).to be_a(Foxtail::CLDR::PatternParser::DateTime::LiteralToken)
+        expect(tokens[5]).to be_a(Foxtail::Intl::PatternParser::DateTime::LiteralToken)
         expect(tokens[5].value).to eq(", ")
 
-        expect(tokens[6]).to be_a(Foxtail::CLDR::PatternParser::DateTime::FieldToken)
+        expect(tokens[6]).to be_a(Foxtail::Intl::PatternParser::DateTime::FieldToken)
         expect(tokens[6].value).to eq("yyyy")
       end
 
@@ -141,26 +141,26 @@ RSpec.describe Foxtail::CLDR::PatternParser::DateTime do
         tokens = parser.parse("EEEE 'at' HH:mm")
         expect(tokens).to have_attributes(size: 7)
 
-        expect(tokens[0]).to be_a(Foxtail::CLDR::PatternParser::DateTime::FieldToken)
+        expect(tokens[0]).to be_a(Foxtail::Intl::PatternParser::DateTime::FieldToken)
         expect(tokens[0].value).to eq("EEEE")
 
-        expect(tokens[1]).to be_a(Foxtail::CLDR::PatternParser::DateTime::LiteralToken)
+        expect(tokens[1]).to be_a(Foxtail::Intl::PatternParser::DateTime::LiteralToken)
         expect(tokens[1].value).to eq(" ")
 
-        expect(tokens[2]).to be_a(Foxtail::CLDR::PatternParser::DateTime::QuotedToken)
+        expect(tokens[2]).to be_a(Foxtail::Intl::PatternParser::DateTime::QuotedToken)
         expect(tokens[2].value).to eq("'at'")
         expect(tokens[2].literal_text).to eq("at")
 
-        expect(tokens[3]).to be_a(Foxtail::CLDR::PatternParser::DateTime::LiteralToken)
+        expect(tokens[3]).to be_a(Foxtail::Intl::PatternParser::DateTime::LiteralToken)
         expect(tokens[3].value).to eq(" ")
 
-        expect(tokens[4]).to be_a(Foxtail::CLDR::PatternParser::DateTime::FieldToken)
+        expect(tokens[4]).to be_a(Foxtail::Intl::PatternParser::DateTime::FieldToken)
         expect(tokens[4].value).to eq("HH")
 
-        expect(tokens[5]).to be_a(Foxtail::CLDR::PatternParser::DateTime::LiteralToken)
+        expect(tokens[5]).to be_a(Foxtail::Intl::PatternParser::DateTime::LiteralToken)
         expect(tokens[5].value).to eq(":")
 
-        expect(tokens[6]).to be_a(Foxtail::CLDR::PatternParser::DateTime::FieldToken)
+        expect(tokens[6]).to be_a(Foxtail::Intl::PatternParser::DateTime::FieldToken)
         expect(tokens[6].value).to eq("mm")
       end
 
@@ -170,10 +170,10 @@ RSpec.describe Foxtail::CLDR::PatternParser::DateTime do
         tokens = parser.parse("Info: dd/MM/yyyy Output: HH:mm")
 
         # Should parse as: ["Info: ", "dd", "/", "MM", "/", "yyyy", " Output: ", "HH", ":", "mm"]
-        expect(tokens[0]).to be_a(Foxtail::CLDR::PatternParser::DateTime::LiteralToken)
+        expect(tokens[0]).to be_a(Foxtail::Intl::PatternParser::DateTime::LiteralToken)
         expect(tokens[0].value).to eq("Info: ")
 
-        expect(tokens[1]).to be_a(Foxtail::CLDR::PatternParser::DateTime::FieldToken)
+        expect(tokens[1]).to be_a(Foxtail::Intl::PatternParser::DateTime::FieldToken)
         expect(tokens[1].value).to eq("dd")
       end
 
@@ -182,25 +182,25 @@ RSpec.describe Foxtail::CLDR::PatternParser::DateTime do
         # The 'd' should be parsed as day field even when surrounded by Japanese characters
         tokens = parser.parse("yyyy年MMMMd日(EEEE)")
 
-        expect(tokens[0]).to be_a(Foxtail::CLDR::PatternParser::DateTime::FieldToken)
+        expect(tokens[0]).to be_a(Foxtail::Intl::PatternParser::DateTime::FieldToken)
         expect(tokens[0].value).to eq("yyyy")
 
-        expect(tokens[1]).to be_a(Foxtail::CLDR::PatternParser::DateTime::LiteralToken)
+        expect(tokens[1]).to be_a(Foxtail::Intl::PatternParser::DateTime::LiteralToken)
         expect(tokens[1].value).to eq("年")
 
-        expect(tokens[2]).to be_a(Foxtail::CLDR::PatternParser::DateTime::FieldToken)
+        expect(tokens[2]).to be_a(Foxtail::Intl::PatternParser::DateTime::FieldToken)
         expect(tokens[2].value).to eq("MMMM")
 
-        expect(tokens[3]).to be_a(Foxtail::CLDR::PatternParser::DateTime::FieldToken)
+        expect(tokens[3]).to be_a(Foxtail::Intl::PatternParser::DateTime::FieldToken)
         expect(tokens[3].value).to eq("d")
 
-        expect(tokens[4]).to be_a(Foxtail::CLDR::PatternParser::DateTime::LiteralToken)
+        expect(tokens[4]).to be_a(Foxtail::Intl::PatternParser::DateTime::LiteralToken)
         expect(tokens[4].value).to eq("日(")
 
-        expect(tokens[5]).to be_a(Foxtail::CLDR::PatternParser::DateTime::FieldToken)
+        expect(tokens[5]).to be_a(Foxtail::Intl::PatternParser::DateTime::FieldToken)
         expect(tokens[5].value).to eq("EEEE")
 
-        expect(tokens[6]).to be_a(Foxtail::CLDR::PatternParser::DateTime::LiteralToken)
+        expect(tokens[6]).to be_a(Foxtail::Intl::PatternParser::DateTime::LiteralToken)
         expect(tokens[6].value).to eq(")")
       end
     end
@@ -209,7 +209,7 @@ RSpec.describe Foxtail::CLDR::PatternParser::DateTime do
       it "handles standalone AM/PM correctly" do
         tokens = parser.parse("a")
         expect(tokens).to have_attributes(size: 1)
-        expect(tokens.first).to be_a(Foxtail::CLDR::PatternParser::DateTime::FieldToken)
+        expect(tokens.first).to be_a(Foxtail::Intl::PatternParser::DateTime::FieldToken)
         expect(tokens.first.value).to eq("a")
       end
 
@@ -217,7 +217,7 @@ RSpec.describe Foxtail::CLDR::PatternParser::DateTime do
         # This should treat letters as part of the word, not as field patterns
         tokens = parser.parse("Info")
         expect(tokens).to have_attributes(size: 1)
-        expect(tokens.first).to be_a(Foxtail::CLDR::PatternParser::DateTime::LiteralToken)
+        expect(tokens.first).to be_a(Foxtail::Intl::PatternParser::DateTime::LiteralToken)
         expect(tokens.first.value).to eq("Info")
       end
 
@@ -225,13 +225,13 @@ RSpec.describe Foxtail::CLDR::PatternParser::DateTime do
         tokens = parser.parse("yyyyMMdd")
         expect(tokens).to have_attributes(size: 3)
 
-        expect(tokens[0]).to be_a(Foxtail::CLDR::PatternParser::DateTime::FieldToken)
+        expect(tokens[0]).to be_a(Foxtail::Intl::PatternParser::DateTime::FieldToken)
         expect(tokens[0].value).to eq("yyyy")
 
-        expect(tokens[1]).to be_a(Foxtail::CLDR::PatternParser::DateTime::FieldToken)
+        expect(tokens[1]).to be_a(Foxtail::Intl::PatternParser::DateTime::FieldToken)
         expect(tokens[1].value).to eq("MM")
 
-        expect(tokens[2]).to be_a(Foxtail::CLDR::PatternParser::DateTime::FieldToken)
+        expect(tokens[2]).to be_a(Foxtail::Intl::PatternParser::DateTime::FieldToken)
         expect(tokens[2].value).to eq("dd")
       end
     end
@@ -240,21 +240,21 @@ RSpec.describe Foxtail::CLDR::PatternParser::DateTime do
   describe "Token classes" do
     describe "Token equality" do
       it "considers tokens equal if same class and value" do
-        token1 = Foxtail::CLDR::PatternParser::DateTime::FieldToken.new("yyyy")
-        token2 = Foxtail::CLDR::PatternParser::DateTime::FieldToken.new("yyyy")
+        token1 = Foxtail::Intl::PatternParser::DateTime::FieldToken.new("yyyy")
+        token2 = Foxtail::Intl::PatternParser::DateTime::FieldToken.new("yyyy")
         expect(token1).to eq(token2)
       end
 
       it "considers tokens different if different class" do
-        field_token = Foxtail::CLDR::PatternParser::DateTime::FieldToken.new("yyyy")
-        literal_token = Foxtail::CLDR::PatternParser::DateTime::LiteralToken.new("yyyy")
+        field_token = Foxtail::Intl::PatternParser::DateTime::FieldToken.new("yyyy")
+        literal_token = Foxtail::Intl::PatternParser::DateTime::LiteralToken.new("yyyy")
         expect(field_token).not_to eq(literal_token)
       end
     end
 
     describe "Token string representation" do
       it "returns value as string" do
-        token = Foxtail::CLDR::PatternParser::DateTime::FieldToken.new("yyyy")
+        token = Foxtail::Intl::PatternParser::DateTime::FieldToken.new("yyyy")
         expect(token.to_s).to eq("yyyy")
       end
     end
